Backend/requirements.txt
fastapi
uvicorn
httpx
supabase
python-dotenv
beautifulsoup4
pandas
openai
modal
PyMuPDF
pillow